Split engine into CPU and GPU
