+++ include/jitk/codegen_util.hpp
@@ -154,130 +154,5 @@ inline std::vector<const LoopB*> find_threaded_blocks(const Block &block, Statis
     return threaded_blocks;
 }
 
-template<typename SelfType>
-void create_monolithic_kernel(SelfType &self, std::map<std::string, bool> kernel_config, const std::vector<Block> &block_list) {
-    using namespace std;
-
-    // When creating a monolithic kernel (all instructions in one shared library), we first combine
-    // the instructions and non-temps from each different block and then we create the kernel
-    vector<InstrPtr> all_instr;
-    set<bh_base *> all_non_temps;
-    bool kernel_is_computing = false;
-    for(const Block &block: block_list) {
-        assert(not block.isInstr());
-        block.getAllInstr(all_instr);
-        block.getLoop().getAllNonTemps(all_non_temps);
-        if (not block.isSystemOnly()) {
-            kernel_is_computing = true;
-        }
-    }
-    // Let's find the arrays that are allocated and freed between blocks in the kernel
-    vector<bh_base*> kernel_temps;
-    set<bh_base*> constructors;
-    for(const InstrPtr &instr: all_instr) {
-        if (instr->constructor) {
-            assert(instr->operand[0].base != NULL);
-            constructors.insert(instr->operand[0].base);
-        } else if (instr->opcode == BH_FREE and util::exist(constructors, instr->operand[0].base)) {
-            kernel_temps.push_back(instr->operand[0].base);
-            all_non_temps.erase(instr->operand[0].base);
-        }
-    }
-
-    // Let's create the symbol table for the kernel
-    const SymbolTable symbols(all_instr, all_non_temps, kernel_config);
-    self.stat.record(symbols);
-
-    // Let's execute the kernel
-    if (kernel_is_computing) { // We can skip this step if the kernel does no computation
-        // Create the constant vector
-        vector<const bh_instruction*> constants;
-        constants.reserve(symbols.constIDs().size());
-        for (const InstrPtr &instr: symbols.constIDs()) {
-            constants.push_back(&(*instr));
-        }
-
-        const auto lookup = self.codegen_cache.get(block_list, symbols);
-        if(lookup.second) {
-            // In debug mode, we check that the cached source code is correct
-            #ifndef NDEBUG
-                stringstream ss;
-                self.write_kernel(block_list, symbols, kernel_temps, ss);
-                if (ss.str().compare(lookup.first) != 0) {
-                    cout << "\nCached source code: \n" << lookup.first;
-                    cout << "\nReal source code: \n" << ss.str();
-                    assert(1 == 2);
-                }
-            #endif
-            self.execute(lookup.first, symbols.getParams(), symbols.offsetStrideViews(), constants);
-        } else {
-            const auto tcodegen = chrono::steady_clock::now();
-            stringstream ss;
-            self.write_kernel(block_list, symbols, kernel_temps, ss);
-            string source = ss.str();
-            self.stat.time_codegen += chrono::steady_clock::now() - tcodegen;
-            self.execute(source, symbols.getParams(), symbols.offsetStrideViews(), constants);
-            self.codegen_cache.insert(std::move(source), block_list, symbols);
-        }
-    }
-
-    // Finally, let's cleanup
-    for(bh_base *base: symbols.getFrees()) {
-        bh_data_free(base);
-    }
-}
-
-template<typename SelfType>
-void create_kernel(SelfType &self, std::map<std::string, bool> kernel_config, const std::vector<Block> &block_list) {
-    using namespace std;
-
-    // When creating a regular kernels (a block-nest per shared library), we create one kernel at a time
-    for(const Block &block: block_list) {
-        assert(not block.isInstr());
-
-        // Let's create the symbol table for the kernel
-        const SymbolTable symbols(block.getAllInstr(), block.getLoop().getAllNonTemps(), kernel_config);
-        self.stat.record(symbols);
-
-        // Let's execute the kernel
-        if (not block.isSystemOnly()) { // We can skip this step if the kernel does no computation
-
-            // Create the constant vector
-            vector<const bh_instruction*> constants;
-            constants.reserve(symbols.constIDs().size());
-            for (const InstrPtr &instr: symbols.constIDs()) {
-                constants.push_back(&(*instr));
-            }
-            const auto lookup = self.codegen_cache.get({ block }, symbols);
-            if(lookup.second) {
-                // In debug mode, we check that the cached source code is correct
-                #ifndef NDEBUG
-                    stringstream ss;
-                    self.write_kernel({block}, symbols, {}, ss);
-                    if (ss.str().compare(lookup.first) != 0) {
-                        cout << "\nCached source code: \n" << lookup.first;
-                        cout << "\nReal source code: \n" << ss.str();
-                        assert(1 == 2);
-                    }
-                #endif
-                self.execute(lookup.first, symbols.getParams(), symbols.offsetStrideViews(), constants);
-            } else {
-                const auto tcodegen = chrono::steady_clock::now();
-                stringstream ss;
-                self.write_kernel({block}, symbols, {}, ss);
-                string source = ss.str();
-                self.stat.time_codegen += chrono::steady_clock::now() - tcodegen;
-                self.execute(source, symbols.getParams(), symbols.offsetStrideViews(), constants);
-                self.codegen_cache.insert(std::move(source), {block}, symbols);
-            }
-        }
-
-        // Finally, let's cleanup
-        for(bh_base *base: symbols.getFrees()) {
-            bh_data_free(base);
-        }
-    }
-}
-
 } // jitk
 } // bohrium
