+++ include/jitk/engines/engine.hpp
@@ -27,20 +27,15 @@ If not, see <http://www.gnu.org/licenses/>.
 #include <bh_instruction.hpp>
 #include <boost/filesystem.hpp>
 
-namespace fs = boost::filesystem;
-
 namespace bohrium {
 namespace jitk {
 
 class Engine {
-  public:
+protected:
     const ConfigParser &config;
     Statistics &stat;
-
     FuseCache fcache;
-
     CodegenCache codegen_cache;
-
     const bool verbose;
 
     // Maximum number of cache files
@@ -61,6 +56,7 @@ class Engine {
     // The hash of the JIT compilation command
     size_t compilation_hash;
 
+public:
     Engine(const ConfigParser &config, Statistics &stat) :
       config(config),
       stat(stat),
@@ -71,7 +67,7 @@ class Engine {
       tmp_dir(get_tmp_path(config)),
       tmp_src_dir(tmp_dir / "src"),
       tmp_bin_dir(tmp_dir / "obj"),
-      cache_bin_dir(config.defaultGet<fs::path>("cache_dir", "")),
+      cache_bin_dir(config.defaultGet<boost::filesystem::path>("cache_dir", "")),
       compilation_hash(0) {
         // Let's make sure that the directories exist
         jitk::create_directories(tmp_src_dir);
@@ -84,183 +80,11 @@ class Engine {
 
     virtual ~Engine() {}
 
-    virtual void execute(const std::string &source,
-                         const std::vector<bh_base*> &non_temps,
-                         const std::vector<const bh_view*> &offset_strides,
-                         const std::vector<const bh_instruction*> &constants) {
-        throw std::runtime_error("execute1 - Abstract method must be implemented!");
-    }
-
-    virtual void execute(const std::string &source,
-                         const std::vector<bh_base*> &non_temps,
-                         const std::vector<const LoopB*> &threaded_blocks,
-                         const std::vector<const bh_view*> &offset_strides,
-                         const std::vector<const bh_instruction*> &constants) {
-        throw std::runtime_error("execute2 - Abstract method must be implemented!");
-    }
-
-    virtual void copyToHost(const std::vector<bh_base*> &bases) {
-        throw std::runtime_error("copyToHost - Abstract method must be implemented!");
-    }
-    virtual void copyToHost(const std::set<bh_base*> &bases) {
-        throw std::runtime_error("copyToHost - Abstract method must be implemented!");
-    }
-
-    virtual void copyToDevice(const std::vector<bh_base*> &base_list) {
-        throw std::runtime_error("copyToDevice - Abstract method must be implemented!");
-    }
-    virtual void copyToDevice(const std::set<bh_base*> &base_list) {
-        throw std::runtime_error("copyToDevice - Abstract method must be implemented!");
-    }
-
+    virtual std::string info() const = 0;
+    virtual const std::string write_type(bh_type dtype) = 0;
     virtual void set_constructor_flag(std::vector<bh_instruction*> &instr_list) = 0;
 
-    virtual void delBuffer(bh_base* &base) {
-        throw std::runtime_error("delBuffer - Abstract method must be implemented!");
-    }
-
-    // TODO: Better union of these two `handle_execution` methods
-    virtual void handle_execution(BhIR *bhir) {
-        throw std::runtime_error("handle_execution - Abstract method must be implemented!");
-    }
-    void handle_execution(component::ComponentImplWithChild &comp, BhIR *bhir) {
-        using namespace std;
-
-        const auto texecution = chrono::steady_clock::now();
-
-        map<string, bool> kernel_config = {
-            { "verbose",        config.defaultGet<bool>("verbose",       false) },
-            { "strides_as_var", config.defaultGet<bool>("strides_as_var", true) },
-            { "index_as_var",   config.defaultGet<bool>("index_as_var",   true) },
-            { "const_as_var",   config.defaultGet<bool>("const_as_var",   true) },
-            { "use_volatile",   config.defaultGet<bool>("use_volatile",  false) }
-        };
-        const uint64_t parallel_threshold = config.defaultGet<uint64_t>("parallel_threshold", 1000);
-
-        // Some statistics
-        stat.record(*bhir);
-
-        // Let's start by cleanup the instructions from the 'bhir'
-        vector<bh_instruction*> instr_list;
-
-        set<bh_base*> frees;
-        instr_list = jitk::remove_non_computed_system_instr(bhir->instr_list, frees);
-
-        // Let's free device buffers and array memory
-        for(bh_base *base: frees) {
-            delBuffer(base);
-            bh_data_free(base);
-        }
-
-        // Set the constructor flag
-        if (config.defaultGet<bool>("array_contraction", true)) {
-            set_constructor_flag(instr_list);
-        } else {
-            for (bh_instruction *instr: instr_list) {
-                instr->constructor = false;
-            }
-        }
-
-        // Let's get the block list
-        // NB: 'avoid_rank0_sweep' is set to true when we have a child to offload to.
-        const vector<jitk::Block> block_list = get_block_list(instr_list, config, fcache, stat, &(comp.child) != nullptr);
-
-        for(const jitk::Block &block: block_list) {
-            assert(not block.isInstr());
-
-            // Let's create the symbol table for the kernel
-            const jitk::SymbolTable symbols(block.getAllInstr(), block.getLoop().getAllNonTemps(), kernel_config);
-            stat.record(symbols);
-
-            // We can skip a lot of steps if the kernel does no computation
-            const bool kernel_is_computing = not block.isSystemOnly();
-
-            // Find the parallel blocks
-            const vector<const jitk::LoopB*> threaded_blocks = find_threaded_blocks(block, stat, parallel_threshold);
-
-            // We might have to offload the execution to the CPU
-            if (threaded_blocks.size() == 0 and kernel_is_computing) {
-                if (kernel_config["verbose"]) {
-                    cout << "Offloading to CPU\n";
-                }
-
-                if (&(comp.child) == nullptr) {
-                    throw runtime_error("handle_execution(): threaded_blocks cannot be empty when child == NULL!");
-                }
-
-                auto toffload = chrono::steady_clock::now();
-
-                // Let's copy all non-temporary to the host
-                copyToHost(symbols.getParams());
-
-                // Let's free device buffers
-                for (bh_base *base: symbols.getFrees()) {
-                    delBuffer(base);
-                }
-
-                // Let's send the kernel instructions to our child
-                vector<bh_instruction> child_instr_list;
-                for (const jitk::InstrPtr &instr: block.getAllInstr()) {
-                    child_instr_list.push_back(*instr);
-                }
-                BhIR tmp_bhir(std::move(child_instr_list), bhir->getSyncs());
-                comp.child.execute(&tmp_bhir);
-                stat.time_offload += chrono::steady_clock::now() - toffload;
-                continue;
-            }
-
-            // Let's execute the kernel
-            if (kernel_is_computing) {
-                // We need a memory buffer on the device for each non-temporary array in the kernel
-                copyToDevice(symbols.getParams());
-
-                // Create the constant vector
-                vector<const bh_instruction*> constants;
-                constants.reserve(symbols.constIDs().size());
-                for (const jitk::InstrPtr &instr: symbols.constIDs()) {
-                    constants.push_back(&(*instr));
-                }
-
-                const auto lookup = codegen_cache.get({ block }, symbols);
-                if(lookup.second) {
-                    // In debug mode, we check that the cached source code is correct
-                    #ifndef NDEBUG
-                        stringstream ss;
-                        write_kernel(block, symbols, threaded_blocks, ss);
-                        if (ss.str().compare(lookup.first) != 0) {
-                            cout << "\nCached source code: \n" << lookup.first;
-                            cout << "\nReal source code: \n" << ss.str();
-                            assert(1 == 2);
-                        }
-                    #endif
-                    execute(lookup.first, symbols.getParams(), threaded_blocks, symbols.offsetStrideViews(), constants);
-                } else {
-                    const auto tcodegen = chrono::steady_clock::now();
-                    stringstream ss;
-                    write_kernel(block, symbols, threaded_blocks, ss);
-                    string source = ss.str();
-                    stat.time_codegen += chrono::steady_clock::now() - tcodegen;
-                    execute(source, symbols.getParams(), threaded_blocks, symbols.offsetStrideViews(), constants);
-                    codegen_cache.insert(std::move(source), { block }, symbols);
-                }
-            }
-
-            // Let's copy sync'ed arrays back to the host
-            copyToHost(bhir->getSyncs());
-
-            // Let's free device buffers
-            for(bh_base *base: symbols.getFrees()) {
-                delBuffer(base);
-            }
-
-            // Finally, let's cleanup
-            for(bh_base *base: symbols.getFrees()) {
-                bh_data_free(base);
-            }
-        }
-        stat.time_total_execution += chrono::steady_clock::now() - texecution;
-    }
-
+protected:
     void write_kernel_function_arguments(const jitk::SymbolTable &symbols,
                                          std::stringstream &ss,
                                          const char *array_type_prefix) {
@@ -522,21 +346,6 @@ class Engine {
         }
     }
 
-    // TODO: Better union of these two `write_kernel` methods
-    virtual void write_kernel(const std::vector<Block> &block_list,
-                              const SymbolTable &symbols,
-                              const std::vector<bh_base*> &kernel_temps,
-                              std::stringstream &ss) {
-        throw std::runtime_error("write_kernel1 - Abstract method must be implemented!");
-    };
-
-    virtual void write_kernel(const Block &block,
-                              const SymbolTable &symbols,
-                              const std::vector<const LoopB*> &threaded_blocks,
-                              std::stringstream &ss) {
-        throw std::runtime_error("write_kernel2 - Abstract method must be implemented!");
-    };
-
     virtual void loop_head_writer(const SymbolTable &symbols,
                                   Scope &scope,
                                   const LoopB &block,
@@ -544,71 +353,7 @@ class Engine {
                                   const std::vector<const LoopB *> &threaded_blocks,
                                   std::stringstream &out) = 0;
 
-    virtual std::string info() const = 0;
-
-    virtual const std::string write_type(bh_type dtype) = 0;
-
-    // TODO: Unify the two `handle_extmethod`. They seem very alike.
-    template <typename T>
-    void handle_extmethod(T &comp, BhIR *bhir) {
-        std::vector<bh_instruction> instr_list;
-
-        for (bh_instruction &instr: bhir->instr_list) {
-            auto ext = comp.extmethods.find(instr.opcode);
-
-            if (ext != comp.extmethods.end()) { // Execute the instructions up until now
-                BhIR b(std::move(instr_list), bhir->getSyncs());
-                comp.execute(&b);
-                instr_list.clear(); // Notice, it is legal to clear a moved vector.
-                const auto texecution = std::chrono::steady_clock::now();
-                ext->second.execute(&instr, nullptr); // Execute the extension method
-                stat.time_ext_method += std::chrono::steady_clock::now() - texecution;
-            } else {
-                instr_list.push_back(instr);
-            }
-        }
-
-        bhir->instr_list = instr_list;
-    }
-
-    template <typename T>
-    void handle_extmethod(T &comp, BhIR *bhir, std::set<bh_opcode> child_extmethods) {
-        std::vector<bh_instruction> instr_list;
-
-        for (bh_instruction &instr: bhir->instr_list) {
-            auto ext = comp.extmethods.find(instr.opcode);
-            auto childext = child_extmethods.find(instr.opcode);
-
-            if (ext != comp.extmethods.end() or childext != child_extmethods.end()) {
-                // Execute the instructions up until now
-                BhIR b(std::move(instr_list), bhir->getSyncs());
-                comp.execute(&b);
-                instr_list.clear(); // Notice, it is legal to clear a moved vector.
-
-                if (ext != comp.extmethods.end()) {
-                    const auto texecution = std::chrono::steady_clock::now();
-                    ext->second.execute(&instr, &*this); // Execute the extension method
-                    stat.time_ext_method += std::chrono::steady_clock::now() - texecution;
-                } else if (childext != child_extmethods.end()) {
-                    // We let the child component execute the instruction
-                    std::set<bh_base *> ext_bases = instr.get_bases();
-
-                    copyToHost(ext_bases);
-
-                    std::vector<bh_instruction> child_instr_list;
-                    child_instr_list.push_back(instr);
-                    b.instr_list = child_instr_list;
-                    comp.child.execute(&b);
-                }
-            } else {
-                instr_list.push_back(instr);
-            }
-        }
-
-        bhir->instr_list = instr_list;
-    }
-
-  private:
+private:
     bool need_to_peel(const std::vector<InstrPtr> &ordered_block_sweeps, const Scope &scope) {
         for (const InstrPtr &instr: ordered_block_sweeps) {
             const bh_view &v = instr->operand[0];
