+++ ve/cuda/engine_cuda.hpp
@@ -28,19 +28,17 @@ If not, see <http://www.gnu.org/licenses/>.
 #include <bh_instruction.hpp>
 #include <bh_component.hpp>
 #include <bh_view.hpp>
-#include <jitk/compiler.hpp>
 #include <jitk/statistics.hpp>
 #include <jitk/codegen_util.hpp>
-#include <boost/filesystem.hpp>
 
-#include <jitk/engine.hpp>
+#include <jitk/engines/engine_gpu.hpp>
 
 #include <cuda.h>
 
 namespace {
     // This will output the proper CUDA error strings
     // in the event that a CUDA host call returns an error
-    #define checkCudaErrors(err) __checkCudaErrors (err, __FILE__, __LINE__)
+    #define checkCudaErrors(err) __checkCudaErrors(err, __FILE__, __LINE__)
 
     inline void __checkCudaErrors(CUresult err, const char *file, const int line) {
         if(CUDA_SUCCESS != err) {
@@ -57,7 +55,7 @@ namespace {
 
 namespace bohrium {
 
-class EngineCUDA : public jitk::Engine {
+class EngineCUDA : public jitk::EngineGPU {
 private:
     // Map of all compiled OpenCL programs
     std::map<uint64_t, CUfunction> _functions;
@@ -71,36 +69,8 @@ private:
     const uint64_t work_group_size_3dx;
     const uint64_t work_group_size_3dy;
     const uint64_t work_group_size_3dz;
-    // OpenCL compile flags
-    const std::string compile_flg;
-    // Default device type
-    const std::string default_device_type;
-    // Default platform number
-    const int platform_no;
-    // Returns the global and local work OpenCL ranges based on the 'threaded_blocks'
-    //std::pair<cl::NDRange, cl::NDRange> NDRanges(const std::vector<const jitk::LoopB*> &threaded_blocks) const;
     // A map of allocated buffers on the device
     std::map<bh_base*, CUdeviceptr> buffers;
-    // Verbose flag
-    const bool verbose;
-    // Maximum number of cache files
-    const int64_t cache_file_max;
-    // Some statistics
-    jitk::Statistics &stat;
-    // Record profiling statistics
-    const bool prof;
-
-    // Path to a temporary directory for the source and object files
-    const boost::filesystem::path tmp_dir;
-
-    // Path to the temporary directory of the source files
-    const boost::filesystem::path tmp_src_dir;
-
-    // Path to the temporary directory of the binary files (e.g. .so or .cubin files)
-    const boost::filesystem::path tmp_bin_dir;
-
-    // Path to the directory of the cached binary files (e.g. .so or .cubin files)
-    const boost::filesystem::path cache_bin_dir;
 
     // The compiler to use when function doesn't exist
     jitk::Compiler compiler;
@@ -109,7 +79,7 @@ private:
     size_t compilation_hash;
 
     // Returns the block and thread sizes based on the 'threaded_blocks'
-    std::pair<std::tuple<uint32_t, uint32_t, uint32_t>, std::tuple<uint32_t, uint32_t, uint32_t> >
+    std::pair<std::tuple<uint32_t, uint32_t, uint32_t>, std::tuple<uint32_t, uint32_t, uint32_t>>
         NDRanges(const std::vector<const jitk::LoopB*> &threaded_blocks) const;
 
     // Return a kernel function based on the given 'source'
@@ -131,6 +101,7 @@ public:
                       const std::vector<const jitk::LoopB*> &threaded_blocks,
                       std::stringstream &ss) override;
 
+
     // Delete a buffer
     void delBuffer(bh_base* &base) override {
         checkCudaErrors(cuMemFree(buffers[base]));
@@ -141,15 +112,13 @@ public:
     template <typename T>
     CUdeviceptr *getBuffer(T &base) {
         if(buffers.find(base) == buffers.end()) {
-            std::vector<T> vec = {base};
-            copyToDevice(vec);
+            copyToDevice({ base });
         }
         return &buffers[base];
     }
 
     // Copy 'bases' to the host (ignoring bases that isn't on the device)
-    template <typename T>
-    void _copyToHost(T &bases) {
+    void copyToHost(const std::set<bh_base*> &bases) override {
         auto tcopy = std::chrono::steady_clock::now();
         // Let's copy sync'ed arrays back to the host
         for(bh_base *base: bases) {
@@ -166,17 +135,9 @@ public:
         }
         stat.time_copy2host += std::chrono::steady_clock::now() - tcopy;
     }
-    // TODO: Can this be done differently?
-    void copyToHost(const std::vector<bh_base*> &bases) override {
-        _copyToHost(bases);
-    }
-    void copyToHost(const std::set<bh_base*> &bases) override {
-        _copyToHost(bases);
-    }
 
     // Copy 'base_list' to the device (ignoring bases that is already on the device)
-    template <typename T>
-    void _copyToDevice(T &base_list) {
+    void copyToDevice(const std::set<bh_base*> &base_list) override {
 
         // Let's update the maximum memory usage on the device
         if (prof) {
@@ -205,19 +166,12 @@ public:
         }
         stat.time_copy2dev += std::chrono::steady_clock::now() - tcopy;
     }
-    // TODO: Can this be done differently?
-    void copyToDevice(const std::vector<bh_base*> &base_list) override {
-        _copyToDevice(base_list);
-    }
-    void copyToDevice(const std::set<bh_base*> &base_list) override {
-        _copyToDevice(base_list);
-    }
 
     // Copy all bases to the host (ignoring bases that isn't on the device)
-    void allBasesToHost() {
-        std::vector<bh_base*> bases_on_device;
+    void allBasesToHost() override {
+        std::set<bh_base*> bases_on_device;
         for(auto &buf_pair: buffers) {
-            bases_on_device.push_back(buf_pair.first);
+            bases_on_device.insert(buf_pair.first);
         }
         copyToHost(bases_on_device);
     }
